requirements.txt
typer[all]
rich
solana
pyzmq
pillow
base58
pycryptodome